chore: remove preact and update eslint config file to allow only single quotes

+++ eslint.config.js
@@ -23,7 +23,7 @@ export default [
             ],
             "@stylistic/quotes" : [
                 "error",
-                "double",
+                "single",
                 {
                     "allowTemplateLiterals" : true,
                     "avoidEscape" : true,
